+++ src/scripts/delete-documents-from-indices/interfaces.ts
@@ -1,6 +1,6 @@
 import { Search_RequestBody } from "@opensearch-project/opensearch/api/index.js";
 
-export default interface DeleteDocumentsFromIndexOptions {
+export default interface DeleteDocumentsFromIndicesOptions {
   indices: string[];
   queryBody?: Search_RequestBody;
 }